Change filename style to match project. Move small separate functions to util file.
