+++ src/core/event/event-manager.js
